backend/requirements.txt
# Django Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0

# Database
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Environment variables
python-dotenv>=1.0.0

# Storage (Cloudflare R2/S3)
django-storages[boto3]>=1.14.0
boto3>=1.26.0

# Email service - Resend
resend>=0.6.0

# Additional utilities
Pillow>=10.0.0  # For image processing
requests>=2.31.0
